refactor: Relocate EntityUtils.kt
