+++ arcade-dimensions/src/main/kotlin/net/casual/arcade/dimensions/utils/DimensionUtils.kt
@@ -13,7 +13,7 @@ import net.casual.arcade.dimensions.mixins.level.MinecraftServerAccessor
 import net.casual.arcade.extensions.utils.getExtension
 import net.casual.arcade.utils.ArcadeUtils
 import net.casual.arcade.utils.math.location.LocationWithLevel.Companion.asLocation
-import net.casual.arcade.utils.teleportTo
+import net.casual.arcade.utils.entity.teleportTo
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerWorldEvents
 import net.minecraft.core.registries.Registries
 import net.minecraft.network.chat.Component
