+++ arcade-events-server/src/main/java/net/casual/arcade/events/server/mixins/ScoreboardMixin.java
@@ -13,9 +13,9 @@ import net.casual.arcade.events.server.entity.EntityTeamJoinEvent;
 import net.casual.arcade.events.server.entity.EntityTeamLeaveEvent;
 import net.casual.arcade.events.server.player.PlayerTeamJoinEvent;
 import net.casual.arcade.events.server.player.PlayerTeamLeaveEvent;
-import net.casual.arcade.utils.EntityUtilsKt;
 import net.casual.arcade.utils.PlayerUtils;
 import net.casual.arcade.utils.StringUtilsKt;
+import net.casual.arcade.utils.entity.EntityUtilsKt;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.ServerScoreboard;
 import net.minecraft.server.level.ServerPlayer;
